@@ -40,6 +40,8 @@
     "styled-components": "4.2.0"
   },
   "dependencies": {
+    "@storybook/addon-docs": "^6.4.18",
+    "@storybook/addon-postcss": "^2.0.0",
     "@tippy.js/react": "^2.1.1",
     "@types/uuid": "^3.4.4",
     "classnames": "^2.2.6",
@@ -58,14 +60,13 @@
     "@ibm/plex": "^2.0.0",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
-    "@storybook/addon-actions": "5.1.11",
-    "@storybook/addon-backgrounds": "5.1.11",
-    "@storybook/addon-info": "5.1.11",
-    "@storybook/addon-knobs": "5.1.11",
-    "@storybook/addon-viewport": "5.1.11",
-    "@storybook/addons": "5.1.11",
-    "@storybook/react": "5.1.11",
-    "@storybook/storybook-deployer": "2.8.1",
+    "@storybook/addon-actions": "6.4.18",
+    "@storybook/addon-controls": "^6.4.18",
+    "@storybook/addon-essentials": "^6.4.18",
+    "@storybook/addon-viewport": "6.4.18",
+    "@storybook/addons": "6.4.18",
+    "@storybook/react": "6.4.18",
+    "@storybook/storybook-deployer": "2.8.10",
     "@types/lodash": "^4.14.118",
     "@types/react": "^16.8.13",
     "@types/react-dates": "^17.1.5",
@@ -77,27 +78,22 @@
     "@types/react-slick": "^0.23.2",
     "@types/react-table": "^6.7.21",
     "@types/react-tabs": "^2.3.0",
-    "@types/storybook-addon-jsx": "5.4.3",
-    "@types/storybook__addon-actions": "3.4.2",
-    "@types/storybook__addon-info": "4.1.1",
-    "@types/storybook__addon-knobs": "5.0.0",
-    "@types/storybook__react": "4.0.1",
-    "@types/styled-components": "^4.0.3",
+    "@types/storybook-addon-jsx": "7.0.3",
     "@types/tinycolor2": "^1.4.1",
     "autoprefixer": "^9.5.0",
-    "awesome-typescript-loader": "^5.2.1",
     "babel-core": "^6.26.3",
     "babel-loader": "^8.0.4",
     "clean-webpack-plugin": "^2.0.1",
     "commit-analyzer-fail-on-no-release": "^1.0.0",
     "cz-conventional-changelog": "3.3.0",
+    "fork-ts-checker-webpack-plugin": "^6.3.2",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^1.3.1",
     "json-loader": "^0.5.7",
     "lodash": "^4.17.15",
     "prettier": "1.17.0",
     "pretty-quick": "^1.10.0",
-    "react": "^16.8.6",
+    "react": "^17.0.2",
     "react-dates": "^20.2.0",
     "react-docgen-typescript-loader": "^3.0.1",
     "react-dom": "^16.8.6",
@@ -115,19 +111,20 @@
     "rollup-plugin-typescript2": "^0.20.1",
     "semantic-release": "^17.1.1",
     "shx": "^0.3.2",
-    "storybook-addon-jsx": "7.1.2",
-    "storybook-addon-styled-component-theme": "1.2.1",
-    "storybook-readme": "^5.0.8",
-    "styled-components": "4.2.0",
+    "storybook-addon-jsx": "7.3.14",
+    "storybook-addon-styled-component-theme": "2.0.0",
+    "storybook-readme": "^5.0.9",
+    "styled-components": "^5.2.1",
     "tinycolor2": "^1.4.1",
+    "tslib": "^2.3.1",
     "tslint": "^5.11.0",
     "tslint-config-airbnb": "^5.11.0",
     "tslint-config-prettier": "^1.18.0",
     "tsutils": "^3.5.0",
     "typeface-roboto": "0.0.54",
-    "typescript": "3.3.x",
+    "typescript": "^4.1.3",
     "uglify-es": "^3.3.9",
-    "webpack": "^4.25.1",
+    "webpack": "^4.44.0",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.14",
     "webpack-merge": "^4.1.4"
