@@ -1,32 +1,37 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { select, boolean } from '@storybook/addon-knobs/react';
-
 import { Alert } from '../';
 
-storiesOf('Alert', module).add(
-  'Alert',
-  () => (
-    <Alert
-      alertStyle={select(
-        'alertStyle',
-        ['success', 'primary', 'secondary', 'danger', 'warning'],
-        'primary',
-      )}
-      lightMode={boolean('lightMode', false)}
-    >
-      Reminder: Sales meeting at Rm 223 in 10 minutes
-    </Alert>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-
-        This is a Alert
-
-        The default theme mode is dark.
-        `,
+export default {
+  title: 'Components/Feedback/Alert',
+  component: Alert,
+  argTypes: {
+    alertStyle: {
+      options: ['success', 'primary', 'secondary', 'danger', 'warning'],
+      control: { type: 'select' },
+    },
+    lightMode: {
+      control: { type: 'boolean' },
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
     },
   },
+  parameters: {
+    docs: {
+      description: {
+        component: 'The default theme mode, for an Alert is dark.',
+      },
+    },
+  },
+};
+
+export const _Alert = args => (
+  <Alert {...args}>Reminder: Sales meeting at Rm 223 in 10 minutes</Alert>
 );
+
+_Alert.args = {
+  alertStyle: 'danger',
+  lightMode: false,
+};
