@@ -1,31 +1,34 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { select, boolean } from '@storybook/addon-knobs/react';
 
 import { Badge } from '../';
 
-storiesOf('Badge', module).add(
-  'Badge',
-  () => (
-    <Badge
-      badgeSize={select('badgeSize', ['sm', 'md', 'lg'], 'md')}
-      badgeStyle={select(
-        'badgeStyle',
-        ['success', 'primary', 'secondary', 'danger', 'warning'],
-        'primary',
-      )}
-      lightMode={boolean('lightMode', false)}
-    >
-      123
-    </Badge>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-
-        This is a badge
-        `,
+export default {
+  title: 'Components/Data Display/Badge',
+  component: Badge,
+  argTypes: {
+    badgeSize: {
+      options: ['sm', 'md', 'lg'],
+      control: 'inline-radio',
+    },
+    badgeStyle: {
+      options: ['success', 'primary', 'secondary', 'danger', 'warning'],
+      control: 'select',
+    },
+    lightMode: {
+      control: 'boolean',
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
     },
   },
-);
+};
+
+export const _Badge = args => <Badge {...args}>123</Badge>;
+
+_Badge.args = {
+  badgeSize: 'md',
+  badgeStyle: 'primary',
+  lightMode: false,
+};
