@@ -1,52 +1,64 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { select } from '@storybook/addon-knobs/react';
 
 import { Button, ButtonGroup } from '../';
 
-storiesOf('ButtonGroup', module).add(
-  'ButtonGroup',
-  () => (
-    <ButtonGroup
-      role={select('role', ['group', 'toolbar'], 'group')}
-      mode={select('mode', ['checkbox', 'radio'], 'checkbox')}
-    >
-      <Button
-        btnStyle="primary"
-        btnSize="md"
-        data-testid="button-1"
-        value="1"
-        onClick={() => console.log('Button 1')}
-      >
-        One
-      </Button>
-      <Button
-        btnStyle="primary"
-        btnSize="md"
-        data-testid="button-2"
-        value="2"
-        onClick={() => console.log('Button 2')}
-      >
-        Two
-      </Button>
-      <Button
-        btnStyle="primary"
-        btnSize="md"
-        data-testid="button-3"
-        value="3"
-        onClick={() => console.log('Button 3')}
-      >
-        Three
-      </Button>
-    </ButtonGroup>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-        This is a ButtonGroup.  Selected values will end up in the
-        ButtonGroup's state.SelectedValues property.
-        `,
+export default {
+  title: 'Components/Interactions/Button',
+  component: ButtonGroup,
+  argTypes: {
+    role: {
+      options: ['group', 'toolbar'],
+      control: {
+        type: 'inline-radio',
+      },
+    },
+    mode: {
+      options: ['checkbox', 'radio'],
+      control: {
+        type: 'inline-radio',
+      },
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
     },
   },
+};
+
+export const _ButtonGroup = args => (
+  <ButtonGroup {...args}>
+    <Button
+      btnStyle="primary"
+      btnSize="md"
+      value="1"
+      data-testid="button-1"
+      onClick={() => console.log('Button 1')}
+    >
+      One
+    </Button>
+    <Button
+      btnStyle="primary"
+      btnSize="md"
+      value="2"
+      data-testid="button-2"
+      onClick={() => console.log('Button 2')}
+    >
+      Two
+    </Button>
+    <Button
+      btnStyle="primary"
+      btnSize="md"
+      value="3"
+      data-testid="button-3"
+      onClick={() => console.log('Button 3')}
+    >
+      Three
+    </Button>
+  </ButtonGroup>
 );
+
+_ButtonGroup.args = {
+  role: 'group',
+  mode: 'checkbox',
+};
