@@ -1,65 +1,102 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { boolean, select, text } from '@storybook/addon-knobs/react';
-import { action } from '@storybook/addon-actions';
-
 import { Checkbox } from '../';
 
-storiesOf('Checkbox', module).add(
-  'Checkbox',
-  () => (
-    <div>
-      <Checkbox
-        id="myInput1"
-        cbSize={select('cbSize', ['sm', 'md', 'lg'], 'md')}
-        checked={boolean('checked (checkbox 1)', true)}
-        data-testid="checkbox-1"
-        defaultChecked={true}
-        disabled={boolean('disabled (checkbox 1)', false)}
-        displayStyle={select('displayStyle', ['inline', 'stacked'], 'inline')}
-        indeterminate={boolean('indeterminate (checkbox 1)', true)}
-        invalid={boolean('invalid', false)}
-        invalidText={text('invalidText', 'A valid value is required')}
-        invalidTextColor={text('invalidTextColor', '')}
-        onChange={action('onChange 1')}
-        value="1"
-      >
-        One
-      </Checkbox>
-      <Checkbox
-        id="myInput2"
-        cbSize={select('cbSize (checkbox 2)', ['sm', 'md', 'lg'], 'md')}
-        checked={boolean('checked (checkbox 2)', true)}
-        data-testid="checkbox-2"
-        defaultChecked={true}
-        disabled={boolean('disabled (checkbox 2)', false)}
-        displayStyle={select('displayStyle', ['inline', 'stacked'], 'inline')}
-        indeterminate={boolean('indeterminate (checkbox 2)', false)}
-        invalid={boolean('invalid', false)}
-        invalidText={text('invalidText', 'A valid value is required')}
-        invalidTextColor={text('invalidTextColor', '')}
-        value="2"
-      >
-        Two
-      </Checkbox>
-    </div>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
+const description = `
+The Checkbox component improves the styling, layout and behavior of default checkbox input HTML element.
+        
+###### Disabled
+Disabled checkbox are supported. The disabled attribute will apply a lighter color to help indicate the input’s state.
+        
+###### Display Style
+By default, any number of checkboxes that are immediate sibling will be *vertically stacked* and appropriately spaced.
+Alternatively, group checkboxes on the same horizontal row by settings the **displayStyle** prop to **inline**
+`;
 
-				The Checkbox component improves the styling, layout and behavior of default checkbox input HTML element.
-
-        ##### Disabled
-        Disabled checkbox are supported. The disabled attribute will apply a lighter color to help indicate the input’s state.
-
-        ##### Display Style
-				By default, any number of checkboxes that are immediate sibling will be *vertically stacked* and appropriately spaced.
-				
-				Alternatively, group checkboxes on the same horizontal row by settings the **displayStyle** prop to **inline**
-
-      `,
+export default {
+  title: 'Components/Interactions/Checkbox',
+  component: Checkbox,
+  argTypes: {
+    cbSize: {
+      options: ['sm', 'md', 'lg'],
+      control: { type: 'inline-radio' },
+    },
+    checked: {
+      control: { type: 'boolean' },
+    },
+    defaultChecked: {
+      control: { type: 'boolean' },
+    },
+    disabled: {
+      control: { type: 'boolean' },
+    },
+    displayStyle: {
+      options: ['inline', 'stacked'],
+      control: { type: 'inline-radio' },
+    },
+    indeterminate: {
+      control: { type: 'boolean' },
+    },
+    invalid: {
+      control: { type: 'boolean' },
+    },
+    invalidText: {
+      control: { type: 'text' },
+    },
+    invalidTextColor: {
+      control: { type: 'color' },
+    },
+    onChange: {
+      action: { type: 'onChange' },
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
+    },
+    value: {
+      control: false,
+    },
+    id: {
+      control: false,
     },
   },
+  parameters: {
+    docs: {
+      description: {
+        component: description,
+      },
+    },
+  },
+};
+
+const _Checkbox = args => (
+  <div>
+    <Checkbox id="myInput1" value="1" data-testid="checkbox-1" {...args}>
+      One
+    </Checkbox>
+  </div>
 );
+
+export const _Indeterminate = _Checkbox.bind({});
+_Indeterminate.args = {
+  cbSize: 'md',
+  checked: true,
+  disabled: false,
+  displayStyle: 'inline',
+  indeterminate: true,
+  invalid: false,
+  invalidText: 'A valid value is required',
+  defaultChecked: true,
+};
+
+export const _Determinate = _Checkbox.bind({});
+_Determinate.args = {
+  cbSize: 'md',
+  checked: true,
+  disabled: false,
+  displayStyle: 'inline',
+  indeterminate: false,
+  invalid: false,
+  invalidText: 'A valid value is required',
+  defaultChecked: true,
+};
