@@ -1,35 +1,55 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { boolean, text } from '@storybook/addon-knobs/react';
-
 import { Collapse } from '../';
 
-storiesOf('Collapse', module).add(
-  'Collapse',
-  () => (
-    <Collapse
-      data-testid="lorem-collapse"
-      isOpen={boolean('isOpen', true)}
-      transition={text(
-        'transition',
-        'height 300ms cubic-bezier(0.4, 0, 0.2, 1)',
-      )}
-    >
-      <p>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
-        tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
-        veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
-        commodo consequat.
-      </p>
-    </Collapse>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-
-        This is a Collapse
-        `,
+export default {
+  title: 'Components/Data Display/Collapse',
+  component: Collapse,
+  argTypes: {
+    isOpen: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    transition: {
+      control: {
+        type: 'text',
+      },
+    },
+    elementType: {
+      control: false,
+    },
+    onInit: {
+      control: false,
+    },
+    collapseHeight: {
+      control: false,
+    },
+    className: {
+      control: false,
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
+    },
+    onChange: {
+      action: false,
     },
   },
+};
+
+export const _Collapse = args => (
+  <Collapse data-testid="lorem-collapse" {...args}>
+    <p>
+      Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+      tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
+      veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea
+      commodo consequat.
+    </p>
+  </Collapse>
 );
+
+_Collapse.args = {
+  isOpen: true,
+  transition: 'height 300ms cubic-bezier(0.4, 0, 0.2, 1)',
+};
