@@ -1,39 +1,87 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { boolean, select } from '@storybook/addon-knobs/react';
-import { action } from '@storybook/addon-actions';
 import { ic_add as icAdd } from 'react-icons-kit/md/ic_add';
 
 import { IconButton } from '../';
 
-storiesOf('IconButton', module).add(
-  'IconButton',
-  () => (
-    <div>
-      <IconButton
-        data-testid="test-button"
-        icon={icAdd}
-        rounded={boolean('rounded', true)}
-        outline={boolean('outline', false)}
-        btnStyle={select(
-          'btnStyle',
-          ['success', 'primary', 'danger', 'warning'],
-          'primary',
-        )}
-        pixelButtonSize={select('pixelButtonSize', [20, 30, 40, 50, 60], 40)}
-        iconSize={select('iconSize', [20, 30, 40, 50, 60], 20)}
-        disabled={boolean('disabled', false)}
-        onClick={action('Icon Button Clicked!')}
-      />
-    </div>
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-
-        This is a IconButton
-        `,
+export default {
+  title: 'Components/Interactions',
+  component: IconButton,
+  argTypes: {
+    rounded: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    outline: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    btnStyle: {
+      control: {
+        type: 'select',
+        options: ['success', 'primary', 'danger', 'warning'],
+      },
+    },
+    pixelButtonSize: {
+      options: [20, 30, 40, 50, 60],
+      control: {
+        type: 'select',
+      },
+    },
+    iconSize: {
+      options: [20, 30, 40, 50, 60],
+      control: {
+        type: 'select',
+      },
+    },
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    onClick: {
+      action: {
+        type: 'onClick',
+      },
+    },
+    theme: {
+      table: {
+        disable: true,
+      },
+    },
+    btnSize: {
+      control: false,
+    },
+    color: {
+      control: false,
+    },
+    icon: {
+      control: false,
+    },
+    backgroundColor: {
+      control: false,
+    },
+    selected: {
+      control: false,
+    },
+    value: {
+      control: false,
     },
   },
+};
+
+export const _IconButton = args => (
+  <div>
+    <IconButton data-testid="icon-button" icon={icAdd} {...args} />
+  </div>
 );
+
+_IconButton.args = {
+  rounded: true,
+  outline: false,
+  btnStyle: 'primary',
+  pixelButtonSize: 40,
+  iconSize: 20,
+  disabled: false,
+};
