@@ -80,9 +80,7 @@ export class Panel extends React.Component<Props, State> {
     const { theme, children, ...props } = this.props;
     return (
       <ThemeProvider theme={(outerTheme: any) => outerTheme || theme}>
-        <PanelWrapper panelStyle={props.panelStyle} {...props}>
-          {children}
-        </PanelWrapper>
+        <PanelWrapper {...props}>{children}</PanelWrapper>
       </ThemeProvider>
     );
   }
