@@ -1,10 +1,10 @@
-import * as React from "react";
-import styled, { ThemeProvider } from "styled-components";
-import { SideNavContext, propsDeepSearch } from "./context";
-import { Themes } from "../themes";
-import { nameSpace } from "../utils/constants";
-import { Icon } from "react-icons-kit";
-import { ic_close } from "react-icons-kit/md/ic_close";
+import * as React from 'react';
+import styled, { ThemeProvider } from 'styled-components';
+import { SideNavContext, propsDeepSearch } from './context';
+import { Themes } from '../themes';
+import { nameSpace } from '../utils/constants';
+import Icon from 'react-icons-kit';
+import { ic_close } from 'react-icons-kit/md/ic_close';
 
 export type Props = {
   /**
@@ -37,7 +37,7 @@ export type Props = {
    * @default false
    **/
   beforeSecondaryToggle?(
-    event: React.MouseEvent<HTMLDivElement, MouseEvent>
+    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
   ): void;
   /**
    * Expand/collapse the secondary sidebar
@@ -51,7 +51,7 @@ export type Props = {
    * @default false
    **/
   afterSecondaryToggle?(
-    event: React.MouseEvent<HTMLDivElement, MouseEvent>
+    event: React.MouseEvent<HTMLDivElement, MouseEvent>,
   ): void;
   /**
    * Handle select SideNavItem
@@ -61,7 +61,7 @@ export type Props = {
   onSelect?(
     event: React.MouseEvent<HTMLDivElement, MouseEvent>,
     selectItemPath: any,
-    itemSecondaryChildren: any
+    itemSecondaryChildren: any,
   ): void;
   /**
    * A  CSS color code
@@ -98,7 +98,7 @@ export type Props = {
    *
    * @default ''
    **/
-  position?: "absolute" | "sticky" | "fixed";
+  position?: 'absolute' | 'sticky' | 'fixed';
   /**
    * Set SideNavbar's distance from top of viewport
    *
@@ -205,7 +205,7 @@ const SSecondarySideNavbar = styled.div`
       : props.theme.sidenav.secondaryNavbar.width)};
   transition: ${(props: any) => props.theme.sidenav.secondaryNavbar.transition};
   visibility: ${(props: any) =>
-    props.isSecondaryNavbarOpen ? "visible" : "hidden"};
+    props.isSecondaryNavbarOpen ? 'visible' : 'hidden'};
   display: flex;
   flex-direction: column;
   .closeIcon {
@@ -237,7 +237,7 @@ export const SideNavbar: React.FunctionComponent<Props> = ({
     beforeToggle,
     onToggle,
     afterToggle,
-    isSecondaryNavbarOpen
+    isSecondaryNavbarOpen,
   };
 
   const handleClick = async (e: any) => {
@@ -252,14 +252,14 @@ export const SideNavbar: React.FunctionComponent<Props> = ({
   // If a child has a SideNav with the 'secondary' prop set to true,
   // its children will automatically be displayed in the secondary SideNavbar
   const activeSideNavItems: any = [];
-  propsDeepSearch(children, "activeSideNavItem", true, activeSideNavItems);
+  propsDeepSearch(children, 'activeSideNavItem', true, activeSideNavItems);
   let activeSideNavItemsChildren: any = [];
   activeSideNavItems.map((child: any) => {
     activeSideNavItemsChildren = [
       ...activeSideNavItemsChildren,
-      ...child.children.filter(
-        (child: any) => (child.props ? child.props.secondary : false)
-      )
+      ...child.children.filter((child: any) =>
+        child.props ? child.props.secondary : false,
+      ),
     ];
   });
 
@@ -278,8 +278,8 @@ export const SideNavbar: React.FunctionComponent<Props> = ({
             itemToggleOpenContent,
             itemToggleCloseContent,
             isSecondaryNavbarOpen,
-            onItemSelect: onSelect
-          }
+            onItemSelect: onSelect,
+          },
         }}
       >
         <SSideNavbar
@@ -295,7 +295,7 @@ export const SideNavbar: React.FunctionComponent<Props> = ({
           {...newProps}
         >
           <a onClick={(e: any) => handleClick(e)} href="#">
-            <Icon className={"closeIcon"} icon={ic_close} />
+            <Icon className={'closeIcon'} icon={ic_close} />
           </a>
           {activeSideNavItemsChildren}
         </SSecondarySideNavbar>
@@ -306,12 +306,12 @@ export const SideNavbar: React.FunctionComponent<Props> = ({
 
 SideNavbar.defaultProps = {
   isOpen: false,
-  width: "",
-  borderLeft: "",
-  borderRight: "",
+  width: '',
+  borderLeft: '',
+  borderRight: '',
   isSecondaryNavbarOpen: false,
-  secondaryNavbarWidth: "",
-  secondaryNavbarHeight: "",
-  itemToggleOpenContent: "",
-  itemToggleCloseContent: ""
+  secondaryNavbarWidth: '',
+  secondaryNavbarHeight: '',
+  itemToggleOpenContent: '',
+  itemToggleCloseContent: '',
 };
