@@ -1,34 +1,89 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
-import { number, boolean, select, text } from '@storybook/addon-knobs/react';
 
 import { Textarea } from '../';
 
-storiesOf('Textarea', module).add(
-  'Textarea',
-  () => (
-    <Textarea
-      id={'myTextarea'}
-      data-testid="textarea"
-      textareaSize={select('textareaSize', ['sm', 'md', 'lg'], 'md')}
-      disabled={boolean('disabled', false)}
-      invalid={boolean('invalid', false)}
-      invalidText={text('invalidText', 'A valid value is required')}
-      invalidTextColor={text('invalidTextColor', 'red')}
-      maxLength={number('maxLength', 1000)}
-      placeholder={text('placeholder', 'Placeholder Text')}
-      required={boolean('required', false)}
-      rows={number('rows', 8)}
-      cols={number('cols', 60)}
-    />
-  ),
-  {
-    info: {
-      text: `
-        ### Notes
-
-        This is a Textarea
-        `,
+export default {
+  title: 'Components/Interactions/Textarea',
+  component: Textarea,
+  argTypes: {
+    theme: {
+      table: {
+        disable: true,
+      },
+    },
+    textareaSize: {
+      control: {
+        type: 'select',
+        options: ['sm', 'md', 'lg'],
+      },
+    },
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    invalid: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    invalidText: {
+      control: {
+        type: 'text',
+      },
+    },
+    invalidTextColor: {
+      control: {
+        type: 'color',
+      },
+    },
+    maxLength: {
+      control: {
+        type: 'number',
+      },
+    },
+    placeholder: {
+      control: {
+        type: 'text',
+      },
+    },
+    required: {
+      control: {
+        type: 'boolean',
+      },
+    },
+    rows: {
+      control: {
+        type: 'number',
+      },
+    },
+    cols: {
+      control: {
+        type: 'number',
+      },
+    },
+    id: {
+      control: false,
+    },
+    value: {
+      control: false,
     },
   },
+};
+
+export const _Textarea = args => (
+  <Textarea id="myTextarea" data-testid="textarea" {...args} />
 );
+
+_Textarea.args = {
+  textareaSize: 'md',
+  disabled: false,
+  invalid: false,
+  invalidText: 'A valid value is required',
+  invalidTextColor: 'red',
+  maxLength: 1000,
+  placeholder: 'Placeholder Text',
+  required: false,
+  rows: 8,
+  cols: 60,
+};
