@@ -1,16 +1,10 @@
 import * as React from 'react';
-import { storiesOf } from '@storybook/react';
 
 import { Typography } from './Typography.component';
 
-storiesOf('Typography', module).add(
-  'Typography',
-  () => (
-    <Typography />
-  ),
-  {
-    info: {
-      disable : true,
-    },
-  },
-);
+export default {
+  title: 'Components/Typography/Typography',
+  component: Typography,
+};
+
+export const _Typography = () => <Typography />;
